% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binit.R
\name{binit}
\alias{binit}
\title{Translated number vector to bin index.}
\usage{
binit(x, xmin = min(x), xmax = max(x), nbin = 100)
}
\arguments{
\item{x}{a numerical vector.}

\item{nbin}{the number of bins.}

\item{xl}{the lower limit of the domain.}

\item{xr}{the upper limit of the domain.}
}
\value{
A list with components
\item{xbin}{a vector of \code{length(nbin)} with elements of the location (bin number) for each x.}
\item{xgrid}{a vector of length(x) representing the interior bin limits.}
\item{nbin}{the number of bins.}
}
\description{
Translates number vector to bin index, given lower and
upper limits of the domain and number of bins. A support function
for histograms.
}
\examples{
binit(runif(20), 0, 1, 5)
}
