% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbase.R
\name{cbase}
\alias{cbase}
\title{Circular B-spline basis}
\usage{
cbase(x, xl = min(x), xr = max(x), nseg = 10, bdeg = 3)
}
\arguments{
\item{x}{a vector of argument values, at which the B-spline basis functions
are to be evaluated.}

\item{xl}{the lower limit of the domain of x.}

\item{xr}{the upper limit of the domain of x.}

\item{nseg}{the number of evenly spaced segments between xl and xr.}

\item{bdeg}{the degree of the basis, usually 1, 2, or 3.}
}
\value{
a matrix with number of rows=\code{length(x)} and
number of columns=\code{nseg}.
}
\description{
Constructs a circular B-spline basis on evenly spaced knots.
}
\examples{
library(JOPS)
xg = seq(0, 4, length = 500)
Bg = cbase(xg, 0, 4, nseg =3, bdeg = 3)
nb1 = ncol(Bg)
matplot(xg, (Bg), type='l',lty=c(1:6), lwd=2, xlab='x', ylab='')

}
\references{
Eilers, P.H.C., Marx, B.D., and Durban, M.
(2015). Twenty years of P-splines, SORT, 39(2): 149-186.
}
\author{
Paul Eilers and Brian Marx
}
