% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdiff.R
\name{cdiff}
\alias{cdiff}
\title{Constructs a second order cyclical differencing matrix}
\usage{
cdiff(n)
}
\arguments{
\item{n}{scalar that determines the dimension of the square differencing matrix.}
}
\value{
a square difference penalty matrix with
\code{n} rows and \code{n} columns.
}
\description{
A difference matrix used for cyclical penalities, where
the ends are penalized to connnect and extrapolation is performed on a cosine curve.
}
\examples{
cdiff(5)

}
\references{
Eilers, P.H.C., Marx, B.D., and Durban, M. (2015).
Twenty years of P-splines, SORT, 39(2): 149-186.
}
\author{
Paul Eilers
}
