% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGHsim.R
\docType{data}
\name{CGHsim}
\alias{CGHsim}
\title{CGHsim Subset}
\format{A dataframe with two columns:
\describe{
  \item{\code{y}}{...}
  \item{\code{x}}{...} }}
\source{

}
\usage{
data(CGHsim)
}
\description{
Simulated data set (subset)
}
\examples{
plot(CGHsim$y, type='b', ylab='y')
}
\references{

}
\keyword{datasets}
