% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Complaints.R
\docType{data}
\name{Complaints}
\alias{Complaints}
\title{Environmental complaints in The Netherlands}
\format{A dataframe with two columns (\code{freq} and \code{count}):
\describe{
  \item{\code{freq}}{The daily number of complaints.}
  \item{\code{count}}{The number of days the specific complaint frequency occurred.}
  }}
\source{

}
\usage{
data(Complaints)
}
\description{
The Rijnomond Environmental Agency registers approximately 20,000
complaints annually from regional inhabitants regarding odors, dust,
noise, among other annoyances. The data are daily number of complaints (
Complaints$freq), ranging from 0-100. The frequency of the
each number of complaints is tallied (Complaints$count). The data
are from 1988.
}
\examples{
plot(Complaints$freq, Complaints$count, type='h',
xlab='Number of complaints per day', ylab='Frequency')
}
\references{

}
\keyword{datasets}
