% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count2d.R
\name{count2d}
\alias{count2d}
\title{Provides the counts in 2D histograms.}
\usage{
count2d(xb, yb, nb)
}
\arguments{
\item{xb}{a vector along \code{x} to be binned in 2D.}

\item{yb}{a vector along \code{y} to be binned in 2D.}

\item{nb}{a vector of length 2 that provides the number of bins for the 2D histogram on \code{x} and \code{y}.}
}
\value{
a matrix of dimension \code{nb[1]} by \code{nb[2]} with counts.
}
\description{
A support function for \code{hist2d()}
that gives the counts in 2D bin.
}
\examples{
count2d(1:4, sample(1:4), c(4,4))
}
