% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disks.R
\docType{data}
\name{Disks}
\alias{Disks}
\title{Prices of hard disk drives}
\format{A dataframe with six columns:
\describe{
  \item{\code{Year}}{1999-2000}
  \item{\code{Month13}}{month of data}
  \item{\code{Size}}{capacity in Gb}
  \item{\code{Buffer}}{buffer size (MB)}
  \item{\code{RPM}}{rotating speed}
  \item{\code{PriceDG}}{in Dutch Guilders, divide by 2.2 for Euro}}}
\source{

}
\usage{
data(Disks)
}
\description{
Prices and capacities of hard disk drives,
as advertised in a Dutch computer monthly in 1999.
Prices are given in Dutch guilders; the Euro did not yet exist.
}
\examples{
plot(Size[Month==5], Price[Month==5], type='b',
xlab='Size', ylab='Price', main='Month of May 1999')
}
\references{

}
\keyword{datasets}
