% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G519C18.R
\docType{data}
\name{G519C18}
\alias{G519C18}
\title{Chromosome G519C18 data}
\format{A dataframe with two columns:
\describe{
  \item{\code{y}}{...}
  \item{\code{x}}{...} }}
\source{
in R: \code{if (!requireNamespace("BiocManager", quietly = TRUE))
install.packages("BiocManager")
BiocManager::install("Vega", version = "3.8")}
}
\usage{
data(G519C18)
}
\description{
Chromosome G519C18 data
}
\examples{
plot(G519C18$x,G519C18$y, type='b', ylab='y', xlab='x')
}
\references{
https://www.bioconductor.org/packages/release/bioc/html/Vega.html
}
\keyword{datasets}
