% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2d.R
\name{hist2d}
\alias{hist2d}
\title{Compile a 2D histogram.}
\usage{
hist2d(x, y, nb = c(100, 100), xlim = c(min(x), max(x)),
  ylim = c(min(y), max(y)))
}
\arguments{
\item{x}{a numeric vector \code{x}.}

\item{y}{a numeric vector \code{y}.}

\item{xlim}{default is min; each is a vector of length 2.}

\item{ylim}{default is max; each is a vector of length 2.}

\item{xb}{a vector of length 2, providing the number of bins along x, y, resp. .}
}
\value{
A list with components
\item{H}{a matrix of dimension \code{nb[1]} by \code{nb[2]} containing bin counts.}
\item{xgrid}{a vector of \code{length(nb[1])} representing the interior bin limits.}
\item{ygrid}{a vector of \code{length(nb[2])} representing the interior bin limits.}
\item{xbin}{a vector of \code{length(x)} with elements of the location (bin number) for each x.}
\item{ybin}{a vector of \code{length(y)} with elements of the location (bin number) for each x.}
}
\description{
Compile a 2D histogram.
}
\examples{
#' data(faithful)
x=faithful$eruptions
y=faithful$waiting
C=hist2d(x,y, c(50,50))
image(C$xgrid,C$ygrid, C$H, xlab='Eruptions', ylab='Wating')

}
