% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2dsm.R
\name{hist2dsm}
\alias{hist2dsm}
\title{Smoothing 2D histograms using P-splines}
\usage{
hist2dsm(Y, nsegx = 10, nsegy = nsegx, bdeg = 3, lambdax = 10,
  lambday = lambdax, dx = 3, dy = dx, Mu.start = NULL,
  kappa = 1e-04)
}
\arguments{
\item{Y}{a matrix of 2D histogram counts.}

\item{nsegx}{the number of evenly spaced knots along x for Tensor product B-spline basis.}

\item{nsegy}{the number of evenly spaced knots along y for Tensor product B-spline basis.}

\item{bdeg}{the degree of the basis, usually 1, 2, or 3.}

\item{lambdax}{the positive number for the tuning parameter along \code{x}.}

\item{lambday}{the positive number for the tuning parameter along \code{y}.}

\item{dx}{the integer (e.g. 1, 2, 3) for the order of the penalty along \code{x}.}

\item{dy}{the integer (e.g. 1, 2, 3) for the order of the penalty along \code{y}.}

\item{kappa}{a (small, positive) number for ridge tuning parameter, if needed.}

\item{ndeg}{the integer (e.g. 1, 2, 3) for the degree of the B-spline tensor basis.}

\item{nseg}{the number of evenly spaced segments between \code{xl} and \code{xr}.}
}
\value{
A list comprising of
\item{ED}{the effective dimension of the smooth 2D surface}
\item{Mu}{a matrix with \code{dim(Y)} consisting of
P-spline density estimates}
}
\description{
Fits 2D smooth P-spline surface using
Poisson response layed out in a 2D histogram
}
\examples{
S=hist2dsm(hist2d(runif(1000),runif(1000),c(100,100))$H, nsegx = 5,
nsegy = 5, bdeg = 3, lambdax = 10, lambday = 10, dx = 3, dy =3,
Mu.start = NULL, kappa = 1e-4)
image(seq(0,1,length=100),seq(0,1,length=100),S$Mu, xlab='x', ylab='y')

}
\references{
Eilers, P.H.C., Marx, B.D., and Durban, M. (2015).
Twenty years of P-splines, SORT, 39(2): 149-186.
}
\author{
Paul Eilers and Brian Marx
}
