% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pclm.R
\name{pclm}
\alias{pclm}
\title{Fittting composite link functions using P-splines}
\usage{
pclm(y, C, X, lambda = 1, pord = 2, itmax = 50, show = F)
}
\arguments{
\item{y}{a vector, usually of counts to be fit.}

\item{C}{a composition matrix with number of rows equal to \code{length(y)} and
number columns equal to the latent distribution, \code{q=length(gamma)}.}

\item{X}{a matrix, ususally a B-spline basis built along the centers of
the bins, with the number of rows \code{q=length(gamma)}. The basis should have enough
columns to have sufficient flexibilty with \code{ncol(X)=n}
 A number that is the upper limit of \code{x}.}

\item{lambda}{the number for the (positive) tuning parameter of the penalty.}

\item{pord}{the number for the order of the difference penalty (default=2) usually 1, 2, or 3.}

\item{itmax}{the maximum number of iterations (defualt=50).}

\item{show}{Set to T or F to display iteration history.}
}
\value{
\item{coeff}{a vector of length \code{n} of estimated P-spline coefficients.}
\item{gamma}{a vector of length \code{q} for the latent coefficients.}
\item{mu}{a vector of length \code{m} of estimated means.}
\item{dev}{Deviance of fit}
\item{ed}{effective dimension}
\item{aic}{AIC}.
}
\description{
pclm is used to fit penalized composite link functions.
It can be useful for density estimation with coarsely aggregated data
or digit prefence data.
}
\examples{
library(ggplot2)
library(JOPS)
cb <- c( 0, 20, 30, 40, 50, 60)
ce <- c(20, 30, 40, 50, 60, 70)
y <- c(79, 54, 19, 1, 1, 0)
m <- length(y)
n <- ce[m]
C <- matrix(0, m, n)
for (i in 1:m) C[i, cb[i]:ce[i]] <- 1
mids = (cb + ce) / 2 - 0.5
widths = ce - cb + 1
dens = y / widths / sum(y)
x=1:n
fit = pclm(y, C, X=bbase(x), lambda = 2, pord = 2, show = T)
gam2 = fit$gamma
#Plot with ggplot2
Data = data.frame(x = mids, y = dens)
Fit = data.frame(x = x - 0.5, y = gam2 / sum(gam2))
Dat = data.frame(cb, ce, dens)
plt = ggplot(Dat, aes(ymin = 0)) +
 geom_rect(aes(xmin = cb, xmax = ce, ymax = dens, fill = I("grey"),color=I('white'))) +
 geom_line(aes(x = x, y = y), size = 1, linetype = 1,data = Fit, col = I("blue")) +
 xlab("Lead concentration") + ylab("Density") +
 ggtitle("Lead in blood") +
 JOPS_theme()
set_window()
print(plt)

}
\references{
Eilers, P. H. C. (2007). III-posed problems with counts, the composite link
model and penalized likelihood. Statistical Modelling, 7(3), 239–254.
}
\author{
Paul Eilers and Jutta Gampe
}
