% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psNormal.R
\name{psNormal}
\alias{psNormal}
\title{Smoothing scattered (normal) data using P-splines}
\usage{
psNormal(x, y, xl = min(x), xr = max(x), nseg = 10, bdeg = 3,
  pord = 2, lambda = 1, wts = NULL)
}
\arguments{
\item{x}{the vector for the continuous regressor of \code{length(y)} and the abcissae of fit.}

\item{y}{the response vector, usually continuous data.}

\item{xl}{the number for the min along \code{x}.}

\item{xr}{the number for the max along \code{x}.}

\item{nseg}{the number of evenly spaced segments between \code{xl} and \code{xr}.}

\item{bdeg}{the number of the degree of the basis, usually 1, 2, or 3 (defalult).}

\item{pord}{the number of the order of the difference penalty, usually 1, 2, or 3 (defalult).}

\item{lambda}{the (positive) number for the tuning parameter for the penalty.}

\item{wts}{the vector of weights, default is 1; 0/1 allowed.}

\item{show}{set to T or F to display iteration history.}
}
\value{
\item{coeff}{a vector of length \code{n} of estimated P-spline coefficients.}
\item{muhat}{a vector of length \code{m} of smooth estimated means.}
\item{wts}{a vector of length \code{m} of weights}
\item{dev}{deviance}
\item{effdim}{estimated effective dimension}
\item{aic}{AIC}
\item{ed.resid}{approximate residual df}
\item{sigma}{root MSE}
\item{cv}{a statistic providing the standard error of leave-one-out prediction.}
\item{nseg, bdeg, pord, lambda}{design parameters.}
\item{xgrid}{gridded x values for plotting.}
\item{ygrid}{gridded fitted mean values for plotting.}
\item{lgrid}{gridded lower 2se values for plotting.}
\item{ugrid}{gridded lower 2se values for plotting.}
}
\description{
psNormal is used to smooth scattered (normal) data using P-splines(with identity link function)
}
\examples{
library(JOPS)
#Extract data
library(MASS)
#Get the data
data(mcycle)
x = mcycle$times
y = mcycle$accel
fit1 = psNormal(x, y, nseg = 20, bdeg = 3, pord = 2, lambda = .8)
names(fit1)
plot(fit1, se=2, xlab="time (ms)", ylab="accel")
}
\references{
Eilers, P.H.C., Marx, B.D., and Durban, M. (2015).
Twenty years of P-splines, SORT, 39(2): 149-186.
}
\author{
Paul Eilers and Brian Marx
}
