% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp2d.R
\name{psp2d}
\alias{psp2d}
\title{Two-dimensional smoothing scattered (normal) data using P-splines.}
\usage{
psp2d(Data, Pars, XYpred)
}
\arguments{
\item{Data}{a matrix of 3 columns \code{x, y, z} of equal length;
the response is \code{z}.}

\item{Pars}{a matrix of 2 rows, where each the first (second) row
sets the P-spline paramters for \code{x (y)}: \code{min max nseg bdeg lambda pdeg}.}

\item{XYpred}{a matrix with two columns \code{(x,y)} that give the coordinates
of (future) prediction; the default is the data locations.}

\item{min(max)}{A scalar for the min(max) along x or y.}

\item{nseg}{the number of evenly spaced segments between min and max.}

\item{bdeg}{a number for the degree of the basis, usually 1, 2, or 3.}

\item{pord}{the number for the order of the difference penalty, usually 1, 2, or 3.}

\item{lambda}{the (positive) tuning parameter for the penalty.}
}
\value{
\item{coef}{a vector of length \code{(Pars[1,3]+Pars[1,4])*(Pars[1,3]+Pars[1,4])}
of (unfolded) estimated P-spline coefficients.}
\item{fit}{a vector of \code{length(y)} of smooth estimated means (at the \code{x,y} locations).}
\item{pred}{a vector of length \code{nrow(XYpred)} of (future) predictions.}
}
\description{
psNormal is used to smooth scattered
(normal) data using P-splines, with aniosotripic penalization of
tensor product B-splines.
}
\examples{
library(SemiPar)
library(fields)
library(spam)
library(JOPS)
# Get the data
data(ethanol)
m = nrow(ethanol)
x = ethanol$C
y = ethanol$E
z = ethanol$NOx
# Set parameters for domain
xlo <- 7
xhi <- 19
ylo <- 0.5
yhi <- 1.25
# Set P-spline parameters, fit and compute surface
xseg <- 10
xdeg <- 3
xpars <- c(xlo, xhi, xseg, xdeg, 3, 1)
yseg <- 10
ydeg <- 3
ypars <- c(ylo, yhi, yseg, ydeg, 3, 1)
Pars1=rbind(xpars,ypars)
resol=100
Xgrid=seq(xlo,xhi,length=resol)
Ygrid=seq(ylo,yhi,length=resol)
XYgrid=as.matrix(expand.grid(seq(xlo,xhi,length=resol),seq(ylo,yhi,length=resol)))
fit=psp2d(cbind(x,y,z),Pars1, XYpred=XYgrid)
image(Xgrid,Ygrid,matrix(fit$pred,resol,resol), xlab='C',ylab='E')
}
\references{
Eilers, P.H.C., Marx, B.D., and Durban, M. (2015).
Twenty years of P-splines, SORT, 39(2): 149-186.
}
\author{
Paul Eilers and Brian Marx
}
