% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp2dU.R
\name{psp2dU}
\alias{psp2dU}
\title{Varying-coefficient penalized signal regression using P-splines.}
\usage{
psp2dU(y, S, S.index, t.var, Pars, ridge.adj = 1e-06, S.pred = NULL,
  t.pred = NULL, y.predicted = NULL, R = 50, x.lab = "X1",
  y.lab = "X2", z.lab = "A.hat", coef.plot = T, image.plot = T,
  se.bands = T, family = "gaussian", link = "default",
  m.binomial = NULL, wts = NULL, r.gamma = NULL, int = T)
}
\arguments{
\item{y}{a response vector of length \code{m}, usually continuous, binary/bimomial or counts.}

\item{S}{a m x p Signal matrix of regressors.}

\item{S.index}{p-vector for index of Signal (e.g. wavelength)}

\item{t.var}{other (indexing) variable in Alpha surface (e.g. temperature)}

\item{Pars}{a matrix with 2 rows, each with P-spline parameters:
\code{min max nseg deg lambda pdeg}, for row and columns.}

\item{ridge.adj}{A ridge penalty tuning parameter (usually set to small value, e.g. 1e-8 to stabilize estimation).}

\item{S.pred}{a matrix of signal inputs for  new predictions.}

\item{y.predicted}{a vector of responses from a cv data set (assoc. with \code{S.pred}).}

\item{R}{resolution of alpha surface (e.g. 20 or 100).}

\item{x.lab, y.lab, z.lab}{character" labels for estimated alpha coefficient surface.}

\item{coef.plot}{set to T or F to display coefficient perspective surface.}

\item{image.plot}{set to T or F to display coefficient image surface.}

\item{se.bands}{set to T or F to produce se surfaces for plots.}

\item{family}{the response distribution, e.g.
\code{"gaussian", "binomial", "poisson", "Gamma"} distribution. Quotes are needed.}

\item{link}{link function (\code{identity, log, sqrt, logit, probit, cloglog, loglog, recipical}). Quotes are needed.}

\item{m.binomial}{a vector of binomial trials having length(y). Default is 1 vector for binomial, NULL otherwise.}

\item{wts}{the weight vector of \code{length(y)}. Deault is 1.}

\item{r.gamma}{a vector of gamma shape parameters. Default is 1 vector for gamma, NULL otherwise.}

\item{int}{set to T or F to include intercept term in linear predictor.}

\item{S}{the signal regressors, which are either "stacked" or "unfolded",
with dimensions (m x p1) x p2 (i.e m matrices each of p1xp2) or
length p1xp2, respectively.}

\item{x.lab(y.lab, z.lab)}{"character" labels for estimated coefficient surface}

\item{t.pred:}{vector t covariates for new predictions}
}
\value{
\item{coef}{a vector of length \code{(Pars[1,3]+Pars[1,4])*(Pars[1,3]+Pars[1,4])}
of estimated P-spline coefficients for tensor surface.}
\item{summary.predicted}{inverse link prediction vectors, and +/- twice se bands.}
\item{deviance}{deviance of fit.}
\item{eff.df}{the approximate effective dimension of fit.}
\item{aic}{AIC}
\item{df.resid}{approx. df resid.}
\item{cv}{leave-one-out standard error prediction (\code{normal, identity}).}
\item{cv.predicted}{standard error prediction for \code{y.predict} (normal, identity).}
\item{Pars}{nseg, bdeg, pord, lambda design parameters}
\item{Dispersion.parm}{estimate of dispersion, Dev/df.resid.}
\item{summary.predicted}{inverse link prediction vectors, and +/- twice se bands.}
\item{eta.predicted}{estimated linear predictor of \code{length(y)}.}
\item{press.mu}{leave-one-out prediction of mean (\code{normal, identity}).}
\item{bin.percent.correct}{percent correct classification based on 0.5 cut-off (\code{binomial}).}
}
\description{
psp2dU is used to regress a (glm) response onto a two-dimensional
 signal such that the signal coefficients can vary over another covariate \code{t}.
 Aniosotripic penalization of tensor product B-splines produces a 2D coefficient surface that
 can be sliced at \code{t}.
}
\examples{
library(fields)
library(JOPS)
psp2dU(y=rnorm(100),S.index=1:500,
S=matrix(rnorm(100*500),100,500),t.var=rnorm(100),
Pars=rbind(c(-3.5,3.5,10,3, 100, 3),c(-3.5,3.5,20,3,.1,2)),
se.bands=F, image.plot=T)
}
\references{
Eilers, P. H. C. and Marx, B. D. (2003). Multivariate calibration with temperature
interaction using two-dimensional penalized signal regression. Chemometrics
and Intellegent Laboratory Systems, 66, 159–174.
}
\author{
Paul Eilers and Brian Marx
}
