% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psPoisson.R
\name{psPoisson}
\alias{psPoisson}
\title{Smoothing scattered Poisson data using P-splines.}
\usage{
psPoisson(x, y, xl = min(x), xr = max(x), nseg = 10, bdeg = 3,
  pord = 2, lambda = 1, wts = NULL, show = F)
}
\arguments{
\item{x}{the vector for the continuous regressor of \code{length(y)} and
the abcissae of fit.}

\item{y}{the response vector, usually count data.}

\item{xl}{the number for the min along \code{x}.}

\item{xr}{the number for the max along \code{x}.}

\item{nseg}{the number of evenly spaced segments between \code{xl} and \code{xr}.}

\item{bdeg}{the number of the degree of the basis, usually 1, 2, or 3 (defalult).}

\item{pord}{the number of the order of the difference penalty, usually 1, 2, or 3. (defalult)}

\item{lambda}{the (positive) number for the tuning parameter for the penalty.}

\item{wts}{the vector of weights, default is 1, 0/1 allowed}

\item{show}{Set to T or F to display iteration history.}
}
\value{
\item{pcoeff}{a vector of length \code{n} of estimated P-spline coefficients.}
\item{muhat}{a vector of length \code{m} of estimated means.}
\item{dev}{Deviance of fit.}
\item{effdim}{effective dimension of fit.}
\item{aic}{AIC}
\item{wts}{a vector of weights.}
\item{nseg, bdeg, pord, lambda}{design parameters.}
\item{xgrid}{gridded x values for plotting.}
\item{ygrid}{gridded fitted mean values for plotting.}
\item{lgrid}{gridded lower 2se values for plotting.}
\item{ugrid}{gridded lower 2se values for plotting.}
}
\description{
psPoisson is used to smooth scattered
Poisson data using P-splines usign a log link function.
}
\examples{
library(JOPS)
library(boot)
# Extract the data
Count=hist(coal$date, breaks=c(1851:1963),plot=F)$counts
Year=c(1851:1962)
xl=min(Year)
xr=max(Year)
# Poisson smoothing
nseg=20
bdeg=3
fit1=psPoisson(Year, Count, xl,xr,nseg,bdeg,pord=2,lambda=1)
names(fit1)
plot(fit1, xlab="Year", ylab="Count", se=2)
}
\references{
Eilers, P.H.C., Marx, B.D., and Durban, M. (2015).
Twenty years of P-splines, SORT, 39(2): 149-186.
}
\author{
Paul Eilers and Brian Marx
}
