% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdw.R
\docType{data}
\name{rdw}
\alias{rdw}
\title{Observations on the widths of red blood cell distributions (RDW)}
\format{A vector.}
\source{
Erasmus University Medical Centre, Rotterdam, The Netherlands
}
\usage{
data(rdw)
}
\description{
Observations on the widths of red blood cell distributions (RDW)
}
\examples{
data(rdw)
hist(rdw, breaks = 20)
}
\keyword{datasets}
