% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowtens.R
\name{rowtens}
\alias{rowtens}
\title{Compute row tensor product of a matrix}
\usage{
rowtens(X, Y = X)
}
\description{
Compute row tensor product of a matrix a
support function for 2-D histogram smoothing.
}
\author{
Paul Eilers
}
