% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_fit.R
\name{signal.fit}
\alias{signal.fit}
\title{Smooth signal (multivariate calibration) regression using P-splines.}
\usage{
signal.fit(response, x.index, x.signal, m.binomial = NULL, nseg = 8,
  bdeg = 3, pord = 3, wts = NULL, link = "default",
  family = "gaussian", r.gamma = NULL, lambda = 0,
  y.predicted = NULL, x.predicted = NULL, ridge.adj = 0, int = T,
  coef.plot = T, se.bands = T)
}
\arguments{
\item{response}{a response vector, usually continuous, binomial or count data.}

\item{x.index}{a vector to of length \code{ncol(X)=p}, associated with the
ordering index of the signal. Default is \code{1:ncol(X)}.}

\item{m.binomial}{a vector of binomial trials having length(y). Default is 1 vector for binomial, NULL otherwise.}

\item{nseg}{the number of evenly spaced segments between \code{xl} and \code{xr}.}

\item{bdeg}{the degree of the basis, usually 1, 2, or 3 (defalult).}

\item{pord}{the order of the difference penalty, usually 1, 2, or 3 (defalult).}

\item{wts}{the weight vector of \code{length(y)}. Deault is 1.}

\item{link}{link function (\code{identity, log, sqrt, logit, probit, cloglog, loglog, recipical}). Quotes are needed.}

\item{family}{the response distribution, e.g.
\code{"gaussian", "binomial", "poisson", "Gamma"} distribution. Quotes are needed.}

\item{r.gamma}{a vector of gamma shape parameters. Default is 1 vector for gamma, NULL otherwise.}

\item{lambda}{the (positive) tuning parameter for the penalty.}

\item{y.predicted}{a vector of responses associated
with \code{x.predicted} which are used to calculated standard error of external prediction. Default is NULL.}

\item{x.predicted}{a matrix of external signals to yield exernal prediction.}

\item{ridge.adj}{A ridge penalty tuning parameter (usually set to small value, e.g. 1e-8 to stabilize estimation).}

\item{int}{set to T or F to include intercept term in linear predictor.}

\item{coef.plot}{set to T or F to display default plots.}

\item{se.bands}{set to T or F to produce se bands in plots.}

\item{X}{a matrix of continuous regressor with nrow(X)=length(y), often
a discrete digitization of a signal or histogram or time series.}
}
\value{
\item{coeff}{a vector with \code{length(n)} of estimated P-spline coefficients.}
\item{mu}{a vector with \code{length(m)} of estimated means.}
\item{eta}{a vector of \code{length(m)} of estimated linear predictors.}
\item{b}{the B-spline basis (for the coefficients), with dimension \code{p}x\code{n}.}
\item{deviance}{deviance of fit.}
\item{eff.df}{the approximate effective dimension of fit.}
\item{aic}{AIC}
\item{df.resid}{approx. df resid.}
\item{summary.beta}{a matrix of dimension \code{px3}, containing beta, and +/- twice se bands for beta.}
\item{cv}{leave-one-out standard error prediction (\code{normal, identity}).}
\item{cv.predicted}{standard error prediction for \code{y.predict} (normal, identity).}
\item{nseg, bdeg, pord, lambda}{design parameters}
\item{Dispersion.parm}{estimate of dispersion, Dev/df.resid.}
\item{summary.predicted}{inverse link prediction vectors, and +/- twice se bands.}
\item{eta.predicted}{estimated linear predictor of \code{length(y)}.}
\item{press.mu}{leave-one-out prediction of mean (\code{normal, identity}).}
\item{bin.percent.correct}{percent correct classification based on 0.5 cut-off (\code{binomial}).}
}
\description{
Smooth signal (multivariate calibration) regression using P-splines.
Support Functions: \code{pspline.fitter()} and \code{pspline.checker()}.
}
\examples{
library(JOPS)
# Get the data
library(fds)
data(nirc)
iindex=nirc$x
X=nirc$y
sel= 50:650 #1200 <= x & x<= 2400
X=X[sel,]
iindex=iindex[sel]
dX=diff(X)
diindex=iindex[-1]
y=as.vector(labc[1,1:40])
oout=23
dX=t(dX[,-oout])
y=y[-oout]
fit1=signal.fit(y,diindex,dX, nseg=25,lambda=.0001,coef.plot=T)
title(main='25 B-spline segments with tuning=0.0001')
names(fit1)

}
\references{
Marx, B.D. and Eilers, P.H.C. (1999). Generalized linear regression for sampled signals and
       curves: A P-spline approach. Technometrics, 41(1): 1-13.
}
\author{
Brian Marx
}
