% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suicide.R
\docType{data}
\name{Suicide}
\alias{Suicide}
\title{Suicide Data Set}
\format{A dataframe with one column: \code{y}}
\source{
https://cran.r-project.org/web/packages/bde/bde.pdf
}
\usage{
data(Suicide)
}
\description{
The dataset comprises lengths (in days) of psychiatric treatment spells for patients used as controls
in a study of suicide risks. The data have been scaled to the interval [0,1] by dividing each data
sample by the maximum value.
Note: Data is also in library(bde), but is scaled to unit interval
}
\examples{

}
\references{
Silverman, B. (1986).
Density Estimation for Statistics and Data Analysis. Chapman & Hall.

Copas, J. B. and Fryer, M. J. (1980). Density estimation and suicide risks in psychiatric treatment.
Journal of the Royal Statistical Society. Series A, 143(2), 167-176
}
\keyword{datasets}
