% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpower.R
\name{tpower}
\alias{tpower}
\title{Compute a truncated power basis}
\usage{
tpower(x, knots, p)
}
\arguments{
\item{x}{a vector on which the basis is calculated.}

\item{knots}{a vector of truncation points (knots) for piecewise polynomials.}

\item{p}{a scalar power for the basis, e.g. p=3 for cubic TPF.}
}
\value{
a TPF basis of degree \code{p}.
}
\description{
Compute a truncated power basis
}
\examples{
library(JOPS)
# Basis  on grid
xg = seq(0, 4, length = 500)
Tg = cbind(tpower(xg,0, p=1), tpower(xg,1, p=1), tpower(xg,2,1))
matplot(xg, Tg, type='l',lty=c(1:6), lwd=2, xlab='x', ylab='',
main='Linear TPF basis (knots: 1, 2, 3)')

}
\author{
Paul Eilers
}
