% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Varstar.R
\docType{data}
\name{Varstar}
\alias{Varstar}
\title{Brightness of a variable star}
\format{A dataframe with eleven columns : \code{V1-V11}
\describe{
\item{V1}{day index}
\item{V2}{brightness}
\item{V3-V11}{Paul Eilers, personal communication}
}}
\usage{
data(Varstar)
}
\description{
Brightness of a variable star
}
\references{
Paul Eilers (personal communication)
}
\keyword{datasets}
