% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Woodsurf.R
\docType{data}
\name{Woodsurf}
\alias{Woodsurf}
\title{Detailed ordered measurements along a grinded piece of wood}
\format{A dataframe with one column: \code{y}.}
\source{

}
\usage{
data(Woodsurf)
}
\description{
Detailed ordered measurements along a grinded piece of wood
}
\examples{
plot(Woodsurf$y, type='b',
xlab='Position', ylab='y')
}
\references{
Pandit, S.M. and Wu, S.M. (1993).
Time Series and System Analysis with Applications.
Krieger Publishing Company.
}
\keyword{datasets}
